api doc과 test 코드 패키지 변경
