+++ src/test/java/com/metacoding/springv2/AuthControllerTest.java
@@ -1,4 +1,4 @@
-package com.metacoding.springv2.web;
+package com.metacoding.springv2;
 
 import static org.hamcrest.Matchers.startsWith;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -14,7 +14,6 @@ import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.metacoding.springv2.MyRestDoc;
 import com.metacoding.springv2.auth.AuthRequest;
 import com.metacoding.springv2.user.User;
