+++ src/test/java/com/metacoding/springv2/BoardControllerTest.java
@@ -1,4 +1,4 @@
-package com.metacoding.springv2.web;
+package com.metacoding.springv2;
 
 import static org.hamcrest.Matchers.hasSize;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -16,7 +16,6 @@ import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.metacoding.springv2.MyRestDoc;
 import com.metacoding.springv2.board.Board;
 import com.metacoding.springv2.board.BoardRequest;
 import com.metacoding.springv2.core.util.JwtUtil;
