--- src/test/java/com/metacoding/springv2/Springv2ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.metacoding.springv2;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Springv2ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
