+++ src/test/java/com/metacoding/springv2/UserControllerTest.java
@@ -1,9 +1,8 @@
-package com.metacoding.springv2.web;
+package com.metacoding.springv2;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,14 +14,13 @@ import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.metacoding.springv2.MyRestDoc;
 import com.metacoding.springv2.core.util.JwtUtil;
-import com.metacoding.springv2.reply.ReplyRequest;
 import com.metacoding.springv2.user.User;
+import com.metacoding.springv2.user.UserRequest;
 
 @Transactional
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
-class ReplyControllerTest extends MyRestDoc {
+class UserControllerTest extends MyRestDoc {
         @Autowired
         private ObjectMapper om;
 
@@ -41,50 +39,49 @@ class ReplyControllerTest extends MyRestDoc {
                 accessToken = JwtUtil.create(testUser);
         }
 
-        @AfterEach
-        void tearDown() {
-                // 테스트 후 정리 작업 (필요시)
-        }
-
-        // 댓글 작성 성공
+        // 회원수정 성공
         @Test
-        public void save_success_test() throws Exception {
+        public void updateUser_success_test() throws Exception {
                 // given
-                ReplyRequest.SaveDTO saveDTO = new ReplyRequest.SaveDTO("test comment", 1);
+                UserRequest.UpdateDTO updateDTO = new UserRequest.UpdateDTO("update@metacoding.com", "12345");
 
-                String requestBody = om.writeValueAsString(saveDTO);
+                String requestBody = om.writeValueAsString(updateDTO);
 
                 // when
                 ResultActions result = mvc.perform(
-                                MockMvcRequestBuilders.post("/api/replies")
+                                MockMvcRequestBuilders.put("/api/users")
                                                 .header("Authorization", accessToken)
                                                 .contentType(MediaType.APPLICATION_JSON)
                                                 .content(requestBody));
+
                 // then
                 result.andExpect(status().isOk())
-                                .andExpect(jsonPath("$.body.id").isNumber())
-                                .andExpect(jsonPath("$.body.comment").value("test comment"))
-                                .andExpect(jsonPath("$.body.userId").value(1))
+                                .andExpect(jsonPath("$.status").value(200))
+                                .andExpect(jsonPath("$.msg").value("성공"))
+                                .andExpect(jsonPath("$.body.id").value(1))
                                 .andExpect(jsonPath("$.body.username").value("ssar"))
-                                .andExpect(jsonPath("$.body.boardId").value(1))
+                                .andExpect(jsonPath("$.body.email").value("update@metacoding.com"))
+                                .andExpect(jsonPath("$.body.roles").value("USER"))
                                 .andDo(MockMvcResultHandlers.print()).andDo(document);
         }
 
-        // 댓글 삭제 성공
+        // 회원조회 성공
         @Test
-        public void deleteById_success_test() throws Exception {
+        public void getUser_success_test() throws Exception {
                 // given
-                Integer replyId = 4;
+                Integer userId = 1;
                 // when
                 ResultActions result = mvc.perform(
-                                MockMvcRequestBuilders.delete("/api/replies/" + replyId)
+                                MockMvcRequestBuilders.get("/api/users/" + userId)
                                                 .header("Authorization", accessToken));
                 // then
                 result.andExpect(status().isOk())
                                 .andExpect(jsonPath("$.status").value(200))
                                 .andExpect(jsonPath("$.msg").value("성공"))
-                                .andExpect(jsonPath("$.body").isEmpty())
+                                .andExpect(jsonPath("$.body.id").value(1))
+                                .andExpect(jsonPath("$.body.username").value("ssar"))
+                                .andExpect(jsonPath("$.body.email").value("ssar@metacoding.com"))
+                                .andExpect(jsonPath("$.body.roles").value("USER"))
                                 .andDo(MockMvcResultHandlers.print()).andDo(document);
         }
-
 }
